+++ extensions/easydict/src/releaseVersion/releaseNote.ts
@@ -2,52 +2,26 @@
  * @author: tisfeng
  * @createTime: 2022-07-03 22:10
  * @lastEditor: tisfeng
- * @lastEditTime: 2022-10-18 11:01
+ * @lastEditTime: 2022-10-26 11:02
  * @fileName: releaseNote.ts
  *
  * Copyright (c) 2022 by tisfeng, All Rights Reserved.
  */
 
 export const releaseNote = `
-## [v2.6.0] - 2022-10-18
-
-本次更新主要支持了[扩展参数](https://developers.raycast.com/information/arguments)。
-
-### ✨ 新功能
-
-- 新增支持了[Arguments](https://developers.raycast.com/information/arguments)。
-- 新增支持了乌克兰等 25 种语言，总计支持 48 种语言。
-
-![arguments-1666060638](https://raw.githubusercontent.com/tisfeng/ImageBed/main/uPic/arguments-1666060638.png)
-
-![hello-1666060655](https://raw.githubusercontent.com/tisfeng/ImageBed/main/uPic/hello-1666060655.png)
-
-![Ukrainian-1666061433](https://raw.githubusercontent.com/tisfeng/ImageBed/main/uPic/Ukrainian-1666061433.png)
+## [v2.6.1] - 2022-10-26
 
 ### 💎 优化
 
-- 优化了多行文本翻译时详情页面的显示效果。
-- 详情页面添加了 \`Copy Text\` 和打开网页查询的动作。
-
-### 🐞 修复
-
-- 修复了一些已知问题。
+- 优化代码。
+- 更新 README。
 
 #### 如果觉得这个扩展还不错，给个 [Star](https://github.com/tisfeng/Raycast-Easydict) ⭐️ 支持一下吧 (^-^)
 
 ---
 
-### ✨ New
-
-- Added support for [Arguments](https://developers.raycast.com/information/arguments).
-- Added support for 25 languages, including Ukrainian, total 48 languages.
-
 ### 💎 Improvement
 
-- Improved the display effect of the details page when translating multiple lines of text.
-- Added \`Copy Text\` and open the web page query action to the details page.
-
-### 🐞 Fixes
-
-- Fixed some known issues.
+- Optimized code.
+- Updated README.
 `;